Functions

FEM machines

Linear deformation

Model reduction types

Modules = [FinEtools, FinEtoolsDeforLinear.DeforModelRedModule]
Private = true
Order = [:function]

Base functionality: Interior

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{T},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {T<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - functionwith the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.infsup_gh — Method

infsup_gh(self::AbstractFEMMDeforLinear, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the norm of the divergence of the displacement.

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.infsup_sh — Method

infsup_sh(self::AbstractFEMMDeforLinear, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the seminorm of the displacement (square root of the sum of the squares of the derivatives of the components of displacement).

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.mass — Method

mass(self::AbstractFEMMDeforLinear,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.stiffness — Method

stiffness(self::AbstractFEMMDeforLinear, assembler::A,
      geom::NodalField{FFlt},
      u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute and assemble stiffness matrix.

Note

The material stiffness matrix is assumed to be the same at all the points of the domain (homogeneous material).

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.thermalstrainloads — Method

thermalstrainloads(self::AbstractFEMMDeforLinear, assembler::A,
    geom::NodalField{FFlt}, u::NodalField{T},
    dT::NodalField{FFlt}) where {A<:AbstractSysvecAssembler, T<:Number}

Compute the thermal-strain load vector.

Base functionality: Boundary

FinEtoolsDeforLinear.FEMMDeforWinklerModule.surfacenormalspringstiffness — Method

surfacenormalspringstiffness(self::FEMMDeforWinkler, assembler::A,
    geom::NodalField{FFlt}, u::NodalField{T},
    springconstant::FFlt, surfacenormal::SurfaceNormal) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the stiffness matrix of surface normal spring.

Rationale: consider continuously distributed springs between the surface of the solid body and the 'ground', in the direction normal to the surface. If the spring coefficient becomes large, we have an approximate method of enforcing the normal displacement to the surface.gas

FinEtoolsDeforLinear.FEMMDeforSurfaceDampingModule.dampingABC — Method

dampingABC(self::FEMMDeforSurfaceDamping, assembler::A,
              geom::NodalField{FFlt}, u::NodalField{T1},
              impedance::T2, surfacenormal::SurfaceNormal) where {A<:AbstractSysmatAssembler, T1<:Number, T2<:Number}

Compute the damping matrix associated with absorbing boundary conditions (ABC) representation of the effect of infinite extent of inviscid fluid next to the surface.

Simple FE models

Advanced FE models

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMAbstractBase,  geom::NodalField{FFlt})

Associate geometry field with the FEMM.

Compute the correction factors to account for the shape of the elements.

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMAbstractBase,  geom::NodalField{FFlt})

Associate geometry field with the FEMM.

Compute the correction factors to account for the shape of the elements.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::FEMMDeforLinear, geom::NodalField{FFlt}, u::NodalField{T}, dT::NodalField{FFlt}, felist::FIntVec, inspector::F, idat, quantity=:Cauchy; context...) where {T<:Number, F<:Function}

Inspect integration point quantities.

Arguments

geom - reference geometry field

u - displacement field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - functionwith the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.stiffness — Method

stiffness(self::AbstractFEMMDeforLinearMS, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute and assemble stiffness matrix.

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.infsup_gh — Method

infsup_gh(self::AbstractFEMMDeforLinearMS, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the norm of the divergence of the displacement.

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.infsup_sh — Method

infsup_sh(self::AbstractFEMMDeforLinearMS, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the seminorm of the displacement (square root of the sum of the squares of the derivatives of the components of displacement).

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.nzebcloadsstiffness — Method

nzebcloadsstiffness(self::AbstractFEMMDeforLinear, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysvecAssembler, T<:Number}

Compute load vector for nonzero EBC for fixed displacement.

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMAbstractBase,  geom::NodalField{FFlt})

Associate geometry field with the FEMM.

Compute the correction factors to account for the shape of the elements.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.stiffness — Method

stiffness(self::AbstractFEMMDeforLinearNICE, assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute and assemble stiffness matrix.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.stiffness — Method

stiffness(self::AbstractFEMMDeforLinearNICE, geom::NodalField{FFlt},  u::NodalField{T}) where {T<:Number}

Compute and assemble stiffness matrix.

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::F,  geom::NodalField{FFlt}) where {F<:FEMMDeforLinearESNICEH8}

Associate geometry field with the FEMM.

Compute the correction factors to account for the shape of the elements.

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::F,  geom::NodalField{FFlt}) where {F<:FEMMDeforLinearESNICET4}

Associate geometry field with the FEMM.

Compute the correction factors to account for the shape of the elements.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinearESNICE, geom::NodalField{FFlt}, u::NodalField{T}, dT::NodalField{FFlt}, felist::FIntVec, inspector::F, idat, quantity=:Cauchy; context...) where {T<:Number, F<:Function}

Inspect integration point quantities.

Arguments

geom - reference geometry field

u - displacement field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - functionwith the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.stiffness — Method

stiffness(self::AbstractFEMMDeforLinearESNICE, assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute and assemble stiffness matrix.

FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule.infsup_gh — Method

infsup_gh(self::AbstractFEMMDeforLinearESNICE, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the norm of the divergence of the displacement.

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule.infsup_sh — Method

infsup_sh(self::AbstractFEMMDeforLinearESNICE, assembler::A, geom::NodalField{FFlt}, u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the matrix to produce the seminorm of the displacement (square root of the sum of the squares of the derivatives of the components of displacement).

This matrix is used in the numerical infsup test (Klaus-Jurgen Bathe, The inf-sup condition and its evaluation for mixed finite element methods, Computers and Structures 79 (2001) 243-252.)

Note

This computation has not been optimized in any way. It can be expected to be inefficient.

Algorithms

Linear deformation

FinEtoolsDeforLinear.AlgoDeforLinearModule.exportdeformation — Method

AlgoDeforLinearModule.exportdeformation(modeldata::FDataDict)

Algorithm for exporting of the deformation for visualization in Paraview.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"geom" = geometry field

"u" = displacement field, or

"us" = array of tuples (name, displacement field)

"postprocessing" = dictionary with values for keys

"boundary_only" = should only the boundary of the regions be rendered? Default is render the interior.

"file" = name of the postprocessing file

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element mmodel machine (mandatory);

Output

modeldata updated with

modeldata["postprocessing"]["exported"] = array of data dictionaries, one for each exported file. The data is stored with the keys:

"file" - names of exported file

"field" - nodal or elemental field

FinEtoolsDeforLinear.AlgoDeforLinearModule.exportmode — Method

AlgoDeforLinearModule.exportmode(modeldata::FDataDict)

Algorithm for exporting of the mmode shape for visualization in Paraview.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"geom" = geometry field

"u" = displacement field

"W" = Computed free-vibration eigenvectors, neigvs columns

"omega" = Computed free-vibration angular frequencies, array of length neigvs

"postprocessing" = dictionary with values for keys

"boundary_only" = should only the boundary of the regions be rendered? Default is render the interior.

"file" = name of the postprocessing file

"mode" = which mode should be visualized?

"component" = which component of the quantity?

"outputcsys" = output coordinate system

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element mmodel machine (mandatory);

Output

modeldata updated with

modeldata["postprocessing"]["exported"] = see exportdeformation()

FinEtoolsDeforLinear.AlgoDeforLinearModule.exportstress — Method

AlgoDeforLinearModule.exportstress(modeldata::FDataDict)

Algorithm for exporting of the stress for visualization in Paraview.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"geom" = geometry field

"u" = displacement field

"postprocessing" = dictionary with values for keys

"boundary_only" = should only the boundary of the regions be rendered? Default is render the interior.

"file" = name of the postprocessing file

"quantity" = quantity to be exported (default :Cauchy)

"component" = which component of the quantity?

"outputcsys" = output coordinate system

"inspectormeth" = inspector method to pass to inspectintegpoints()

"extrap" = method for extrapolating from the quadrature points to the nodes within one element

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element mmodel machine (mandatory);

Output

modeldata updated with

modeldata["postprocessing"]["exported"] = array of data dictionaries, one for each exported file. The data is stored with the keys:

"file" - name of exported file

"field" - nodal field

FinEtoolsDeforLinear.AlgoDeforLinearModule.exportstresselementwise — Method

AlgoDeforLinearModule.exportstresselementwise(modeldata::FDataDict)

Algorithm for exporting of the elementwise stress for visualization in Paraview.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"geom" = geometry field

"u" = displacement field

"postprocessing" = dictionary with values for keys

"boundary_only" = should only the boundary of the regions be rendered? Default is render the interior.

"file" = name of the postprocessing file

"quantity" = quantity to be exported (default :Cauchy)

"component" = which component of the quantity?

"outputcsys" = output coordinate system

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element mmodel machine (mandatory);

Output

modeldata updated with

modeldata["postprocessing"]["exported"] = array of data dictionaries, one for each exported file. The data is stored with the keys:

"file" - name of exported file

"field" - elemental field

FinEtoolsDeforLinear.AlgoDeforLinearModule.linearstatics — Method

AlgoDeforLinearModule.linearstatics(modeldata::FDataDict)

Algorithm for static linear deformation (stress) analysis.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"essential_bcs" = array of essential boundary condition dictionaries

"traction_bcs" = array of traction boundary condition dictionaries

"temperature_change" = dictionary of data for temperature change

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element model machine (mandatory);

For essential boundary conditions (optional) each dictionary would hold

"displacement" = fixed (prescribed) displacement (scalar), or a function with signature function w = f(x). If this value is not given, zero displacement is assumed.

"component" = which component is prescribed (1, 2, 3)?

"node_list" = list of nodes on the boundary to which the condition applies (mandatory)

For traction boundary conditions (optional) each dictionary would hold key-value pairs

"femm" = finite element model machine (mandatory);

"traction_vector" = traction vector, either a constant numerical vector, or a function to be used to construct a ForceIntensity object, or it could be the ForceIntensity object itself.

Output

modeldata = the dictionary on input is augmented with the keys

"geom" = the nodal field that is the geometry

"u" = the nodal field that is the computed displacement

"temp" = the nodal field that is the temperature change

"work" = work of the applied loads

"timing" = dictionary with timing results

FinEtoolsDeforLinear.AlgoDeforLinearModule.modal — Method

AlgoDeforLinearModule.modal(modeldata::FDataDict)

Modal (free-vibration) analysis solver.

Argument

modeldata = dictionary with values for keys

"fens" = finite element node set

"regions" = array of region dictionaries

"essential_bcs" = array of essential boundary condition dictionaries

For each region (connected piece of the domain made of a particular material), mandatory, the region dictionary contains values for keys:

"femm" = finite element mmodel machine (mandatory);

For essential boundary conditions (optional) each dictionary would hold

"displacement" = fixed (prescribed) displacement (scalar): only zero displacement is allowed for modal analysis.

"component" = which component is prescribed (1, 2, 3)?

"node_list" = list of nodes on the boundary to which the condition applies (mandatory)

Control parameters:

"neigvs" = number of eigenvalues/eigenvectors to compute

"omega_shift"= angular frequency shift for mass shifting

"use_lumped_mass" = true or false? (Default is false: consistent mass)

Output

modeldata= the dictionary on input is augmented with

"geom" = the nodal field that is the geometry

"u" = the nodal field that is the computed displacement

"neigvs" = Number of computed eigenvectors

"W" = Computed eigenvectors, neigvs columns

"omega" = Computed angular frequencies, array of length neigvs

"raw_eigenvalues" = Raw computed eigenvalues

Material models

Material for deformation, base functionality

FinEtoolsDeforLinear.MatDeforModule.dett — Method

dett(::Type{DeforModelRed2DStrain},  C::FMat{T}) where {T}

Compute the determinant of a general square matrix.

FinEtoolsDeforLinear.MatDeforModule.dett — Method

dett(::Type{DeforModelRed3D},  C::FMat{T}) where {T}

Compute the determinant of a general square matrix.

FinEtoolsDeforLinear.MatDeforModule.rotstressvec! — Method

rotstressvec!(::Type{DeforModelRed1D},  outstress::FVec{T},  instress::FVec{T},  Rm::_RotationMatrix) where {T}

Rotate the stress vector by the supplied rotation matrix.

Calculate the rotation of the stress vector to the 'bar' coordinate system given by the columns of the rotation matrix Rm.

outstress = output stress vector, overwritten inside

instress = input stress vector

Rm = columns are components of 'bar' basis vectors on the 'plain' basis vectors

FinEtoolsDeforLinear.MatDeforModule.rotstressvec! — Method

rotstressvec!(::Type{DeforModelRed2DAxisymm},  outstress::FVec{T},  instress::FVec{T},  Rm::_RotationMatrix) where {T}

Rotate the stress vector by the supplied rotation matrix.

Calculate the rotation of the stress vector to the 'bar' coordinate system given by the columns of the rotation matrix Rm.

outstress = output stress vector, overwritten inside

instress = input stress vector

Rm = columns are components of 'bar' basis vectors on the 'plain' basis vectors

FinEtoolsDeforLinear.MatDeforModule.rotstressvec! — Method

rotstressvec!(::Type{DeforModelRed2DStrain},  outstress::FVec{T},  instress::FVec{T},  Rm::_RotationMatrix) where {T}

Rotate the stress vector by the supplied rotation matrix.

Calculate the rotation of the stress vector to the 'bar' coordinate system given by the columns of the rotation matrix Rm.

outstress = output stress vector, overwritten inside

instress = input stress vector

Rm = columns are components of 'bar' basis vectors on the 'plain' basis vectors

FinEtoolsDeforLinear.MatDeforModule.rotstressvec! — Method

rotstressvec!(::Type{DeforModelRed2DStress},  outstress::FVec{T},  instress::FVec{T},  Rm::_RotationMatrix) where {T}

Rotate the stress vector by the supplied rotation matrix.

Calculate the rotation of the stress vector to the 'bar' coordinate system given by the columns of the rotation matrix Rm.

outstress = output stress vector, overwritten inside

instress = input stress vector

Rm = columns are components of 'bar' basis vectors on the 'plain' basis vectors

FinEtoolsDeforLinear.MatDeforModule.rotstressvec! — Method

rotstressvec!(::Type{DeforModelRed3D},  outstress::FVec{T},  instress::FVec{T}, Rm::_RotationMatrix) where {T}

Rotate the stress vector by the supplied rotation matrix.

Calculate the rotation of the stress vector to the 'bar' coordinate system given by the columns of the rotation matrix Rm.

outstress = output stress vector, overwritten inside

instress = input stress vector

Rm = columns are components of 'bar' basis vectors on the 'plain' basis vectors

FinEtoolsDeforLinear.MatDeforModule.strainttov! — Method

strainttov!(::Type{DeforModelRed2DStrain}, v::FVec{T}, t::FMat{T}) where {T}

Convert a symmetric matrix of 2x2 strain components into a 3-component vector.

FinEtoolsDeforLinear.MatDeforModule.strainttov! — Method

strainttov!(::Type{DeforModelRed3D}, v::FVec{T}, t::FMat{T}) where {T}

Convert a symmetric matrix of 3x3 strain components into a 6-component vector.

FinEtoolsDeforLinear.MatDeforModule.strainvdet — Method

strainvdet(::Type{DeforModelRed2DStrain},  Cv::FVec{T}) where {T}

Compute the determinant of a symmetric strain-like square matrix represented as a vector. Remember that the shear strain components are twice the entries of the matrix representation.

FinEtoolsDeforLinear.MatDeforModule.strainvdet — Method

strainvdet(::Type{DeforModelRed3D},  Cv::FVec{T}) where {T}

Compute the determinant of a symmetric strain-like square matrix represented as a vector. Remember that the shear strain components are twice the entries of the matrix representation.

FinEtoolsDeforLinear.MatDeforModule.strainvtot! — Method

strainvtot!(::Type{DeforModelRed2DStrain}, t::FMat{T}, v::FVec{T}) where {T}

Convert a strain 3-vector to a matrix of 2x2 strain components (symmetric tensor).

FinEtoolsDeforLinear.MatDeforModule.strainvtot! — Method

strainvtot!(::Type{DeforModelRed3D}, t::FMat{T}, v::FVec{T}) where {T}

Convert a strain 3-vector to a matrix of 2x2 strain components (symmetric tensor).

FinEtoolsDeforLinear.MatDeforModule.strainvtr — Method

strainvtr(::Type{DeforModelRed2DStrain},  Cv::FVec{T}) where {T}

Compute the trace of a symmetric strain-like square matrix represented as a vector.

FinEtoolsDeforLinear.MatDeforModule.strainvtr — Method

strainvtr(::Type{DeforModelRed3D},  Cv::FVec{T}) where {T}

Compute the trace of a symmetric strain-like square matrix represented as a vector.

FinEtoolsDeforLinear.MatDeforModule.stressttov! — Method

stressttov!(::Type{DeforModelRed2DStrain}, v::FVec{T}, t::FMat{T}) where {T}

Convert a symmetric matrix of 2x2 stress components to a 3-component vector.

FinEtoolsDeforLinear.MatDeforModule.stressttov! — Method

stressttov!(::Type{DeforModelRed2DStress}, v::FVec{T}, t::FMat{T}) where {T}

Convert a symmetric matrix of 2x2 stress components to a 3-component vector.

FinEtoolsDeforLinear.MatDeforModule.stressttov! — Method

stressttov!(::Type{DeforModelRed3D}, v::FVec{T}, t::FMat{T}) where {T}

Convert a symmetric matrix of 3x3 stress components to a 6-component vector.

FinEtoolsDeforLinear.MatDeforModule.stressvtot! — Method

stressvtot!(::Type{DeforModelRed2DAxisymm}, t::FMat{T}, v::FVec{T}) where {T}

Convert a 4-vector to a matrix of 3x3 stress components (tensor).

Convert a 4-vector to a symmetric matrix of 3x3 stress components (tensor).

The stress vector components need to be ordered as: sigmax, sigmay, sigmaz, tauxy.

FinEtoolsDeforLinear.MatDeforModule.stressvtot! — Method

stressvtot!(::Type{DeforModelRed2DStrain}, t::FMat{T}, v::FVec{T}) where {T}

Convert a vector to a matrix of 2x2 stress components (symmetric tensor).

If v has 4 entries, also the t[3,3] matrix entry is set.

The stress vector components need to be ordered as: sigmax, sigmay, tauxy, sigmaz, which is the ordering used for the plane-strain model reduction.

FinEtoolsDeforLinear.MatDeforModule.stressvtot! — Method

stressvtot!(::Type{DeforModelRed2DStress}, t::FMat{T}, v::FVec{T}) where {T}

Convert a 3-vector to a matrix of 2x2 stress components (symmetric tensor).

FinEtoolsDeforLinear.MatDeforModule.stressvtot! — Method

stressvtot!(::Type{DeforModelRed3D}, t::FMat{T}, v::FVec{T}) where {T}

Convert a 6-vector to a matrix of 3x3 stress components (symmetric tensor).

FinEtoolsDeforLinear.MatDeforModule.tens4checksymmetry — Method

tens4checksymmetry(C4th)

If the fourth-order tensor of material elasticity has the full set of symmetries, return true; otherwise false.

FinEtoolsDeforLinear.MatDeforModule.tens4deviator! — Method

tens4deviator!(t::Array{T, 4}) where {T}

Compute 4th-order deviator tensor.

Double contraction of a second order tensor with this fourth-order tensor produces the deviator part of the second order tensor.

Example

The product of the deviator tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4deviator!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show tr((S - tr(S)/3*I) ), tr(tS)

FinEtoolsDeforLinear.MatDeforModule.tens4dot2! — Method

tens4dot2!(R::Array{T, 2}, F::Array{T, 4}, S::Array{T, 2}) where {T}

Compute the double contraction of a 4th-order and a 2nd-order tensors.

Note

The double contraction of two second-order sensors is defined as A:B = tr(A'*B) = A_ij B_ij

The resulting second-order tensor is first zeroed out, and then the result is accumulated.

FinEtoolsDeforLinear.MatDeforModule.tens4identity! — Method

tens4identity!(t::Array{T, 4}) where {T}

Compute 4th-order identity tensor.

Example

The product of the identity tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4identity!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show S - tS

FinEtoolsDeforLinear.MatDeforModule.tens4ijkl! — Method

tens4ijkl!(t::Array{T, 4}, A::FA, B::FB, op = :+) where {T, FA, FB}

Fill a 4th-order tensor as a dyadic product of two 2nd-order tensors.

The i,j,k,l component is given as t[i,j,k,l]=A(i,j)*B(k,l).

Note

The tensor is accumulated to. It needs to be initialized to zero, if that is desired as the initial state.

Example

t = fill(0.0, 3, 3, 3, 3)
delta = (I, J) -> I == J ? 1.0 : 0.0
tens4ijkl!(t, delta, delta)
S = rand(3, 3)
@show tr(S) * I
tS = fill(0.0, 3, 3)
@show tens4dot2!(tS, t, S)

FinEtoolsDeforLinear.MatDeforModule.tens4ikjl! — Method

tens4ikjl!(t::Array{T, 4}, A::FA, B::FB) where {T, FA, FB}

Fill a 4th-order tensor as a dyadic product of two 2nd-order tensors.

The i,j,k,l component is given as t[i,j,k,l]=A(i,k)*B(j,l).

Note

The tensor is accumulated to. It needs to be initialized to zero, if that is desired as the initial state.

Example

t = fill(0.0, 3, 3, 3, 3)
delta = (I, J) -> I == J ? 1.0 : 0.0
tens4ikjl!(t, delta, delta)
S = rand(3, 3)
@show transpose(S) 
tS = fill(0.0, 3, 3)
@show transpose(S) - tens4dot2!(tS, t, S)

FinEtoolsDeforLinear.MatDeforModule.tens4iljk! — Method

tens4iljk!(t::Array{T, 4}, A::FA, B::FB) where {T, FA, FB}

Fill a 4th-order tensor as a dyadic product of two 2nd-order tensors.

The i,j,k,l component is given as t[i,j,k,l]=A(i,l)*B(j,k).

Note

The tensor is accumulated to. It needs to be initialized to zero, if that is desired as the initial state.

Example

t = fill(0.0, 3, 3, 3, 3)
delta = (I, J) -> I == J ? 1.0 : 0.0
tens4iljk!(t, delta, delta)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
@show S - tens4dot2!(tS, t, S)

FinEtoolsDeforLinear.MatDeforModule.tens4skewor! — Method

tens4skewor!(t::Array{T, 4}) where {T}

Compute 4th-order skewor tensor.

Double contraction of a second order tensor with this fourth-order tensor produces the skew part of the second order tensor.

Example

The product of the skewor tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4skewor!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show (S - S')/2 * I - tS

FinEtoolsDeforLinear.MatDeforModule.tens4symmetrizor! — Method

tens4symmetrizor!(t::Array{T, 4}) where {T}

Compute 4th-order symmetrizor tensor.

Double contraction of a second order tensor with this fourth-order tensor produces the symmetric part of the second order tensor.

Example

The product of the symmetrizor tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4symmetrizor!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show (S + S')/2 * I - tS

FinEtoolsDeforLinear.MatDeforModule.tens4symmt6x6tot! — Method

tens4symmt6x6tot!(ST::Array{T, 4}, M::FMat{T}) where {T}

Convert a symmetric 6 x 6 matrix to a symmetric 4th-order tensor.

!!! Note The order corresponds to the arrangement of the components of stress (or strain) tensor, symmetric, three-dimensional, into a 6-component vector.

FinEtoolsDeforLinear.MatDeforModule.tens4symmtto6x6t! — Method

tens4symmtto6x6t!(M::FMat{T}, ST::Array{T, 4}) where {T}

Convert a symmetric 4th-order tensor to a 6 x 6 matrix.

!!! Note The order corresponds to the arrangement of the components of stress (or strain) tensor, symmetric, three-dimensional, into a 6-component vector.

Example

J=tens4_ijkl(eye(3),eye(3))
produces the tracor:
T=rand(3); 
sum(diag(T))*eye(3)
t= tens4_dot_2(J,T)
M= tens4_symm_to_6(ST)

FinEtoolsDeforLinear.MatDeforModule.tens4tracor! — Method

tens4tracor!(t::Array{T, 4}) where {T}

Compute 4th-order tracor tensor.

Double contraction of a second order tensor with this fourth-order tensor produces the spherical part of the second order tensor.

Example

The product of the tracor tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4tracor!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show tr(S) * I - tS

FinEtoolsDeforLinear.MatDeforModule.tens4transposor! — Method

tens4transposor!(t::Array{T, 4}) where {T}

Compute 4th-order transposor tensor.

Example

The product of the transposor tensor with the second-order tensor S is

t = fill(0.0, 3, 3, 3, 3)
tens4transposor!(t)
S = rand(3, 3)
tS = fill(0.0, 3, 3)
tens4dot2!(tS, t, S)
@show S' - tS

Material models for elasticity

FinEtoolsDeforLinear.MatDeforLinearElasticModule.tangentmoduli! — Method

tangentmoduli!(self::AbstractMatDeforLinearElastic,  D::FFltMat,  t::FFlt, dt::FFlt, loc::FFltMat, label::FInt)

Calculate the material stiffness matrix.

D = matrix of tangent moduli, supplied as a buffer and overwritten. Returned

as output.

FinEtoolsDeforLinear.MatDeforLinearElasticModule.thermalstrain! — Function

thermalstrain!(self::AbstractMatDeforLinearElastic, thstrain::FFltVec, dT= 0.0)

Compute thermal strain from the supplied temperature increment.

thstrain = thermal strain vector, supplied as buffer, returned as output.

FinEtoolsDeforLinear.MatDeforLinearElasticModule.update! — Function

update!(self::AbstractMatDeforLinearElastic,  stress::FFltVec, output::FFltVec,  strain::FFltVec, thstrain::FFltVec=zeros(6), t::FFlt= 0.0, dt::FFlt= 0.0,  loc::FFltMat=zeros(3,1), label::FInt=0, quantity=:nothing)

Update material state.

strain = strain vector,

thstrain = thermal strain vector,

t = current time,

dt = current time step,

loc = location of the quadrature point in global Cartesian coordinates,

label = label of the finite element in which the quadrature point is found.

Output

stress = stress vector, allocated by the caller with a size of the number of stress and

strain components, nstressstrain. The components of the stress vector are calculated and stored in the stress vector.

output = array which is (if necessary) allocated in an appropriate size, filled with the output quantity, and returned.