Types

Types

FEM machines

Linear deformation

FinEtoolsDeforLinear.DeforModelRedModule.AbstractDeforModelRed — Type.

AbstractDeforModelRed

Abstract type of model reduction for deformation analysis.

source

FinEtoolsDeforLinear.DeforModelRedModule.DeforModelRed1D — Type.

DeforModelRed1D <: AbstractDeforModelRed

Model reduction for 1-D models.

source

FinEtoolsDeforLinear.DeforModelRedModule.DeforModelRed2DAxisymm — Type.

DeforModelRed2DAxisymm <: AbstractDeforModelRed

Model reduction for 2-D axially symmetric models.

source

FinEtoolsDeforLinear.DeforModelRedModule.DeforModelRed2DStrain — Type.

DeforModelRed2DStrain <: AbstractDeforModelRed

Model reduction for 2-D plane strain models.

source

FinEtoolsDeforLinear.DeforModelRedModule.DeforModelRed2DStress — Type.

DeforModelRed2DStress <: AbstractDeforModelRed

Model reduction for 2-D plane stress models.

source

FinEtoolsDeforLinear.DeforModelRedModule.DeforModelRed3D — Type.

DeforModelRed3D <: AbstractDeforModelRed

Model reduction for 3-D models.

source

FinEtoolsDeforLinear.FEMMDeforLinearBaseModule.AbstractFEMMDeforLinear — Type.

AbstractFEMMDeforLinear <: AbstractFEMMBase

Abstract type of FEMM for linear deformation.

source

FinEtoolsDeforLinear.FEMMDeforLinearModule.FEMMDeforLinear — Type.

FEMMDeforLinear{MR<:AbstractDeforModelRed,  S<:AbstractFESet, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinear

Class for linear deformation finite element modeling machine.

source

FinEtoolsDeforLinear.FEMMDeforLinearModule.FEMMDeforLinear — Method.

FEMMDeforLinear(mr::Type{MR}, integdomain::IntegDomain{S, F}, material::M) where {MR<:AbstractDeforModelRed, S<:AbstractFESet, F<:Function, M<:AbstractMatDeforLinearElastic}

Constructor of linear deformation finite element modeling machine.

source

FinEtoolsDeforLinear.FEMMDeforWinklerModule.FEMMDeforWinkler — Type.

FEMMDeforWinkler{S<:AbstractFESet, F<:Function} <: AbstractFEMM

Type for normal spring support (Winkler).

source

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.AbstractFEMMDeforLinearMS — Type.

AbstractFEMMDeforLinearMS <: AbstractFEMMDeforLinear

Abstract type for mean-strain linear deformation FEMM.

source

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.FEMMDeforLinearMSH8 — Type.

FEMMDeforLinearMSH8{MR<:AbstractDeforModelRed, S<:FESetH8, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearMS

Type for mean-strain linear deformation FEMM based on eight-node hexahedral elements.

source

FinEtoolsDeforLinear.FEMMDeforLinearMSModule.FEMMDeforLinearMST10 — Type.

FEMMDeforLinearMST10{MR<:AbstractDeforModelRed, S<:FESetT10, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearMS

Type for mean-strain linear deformation FEMM based on 10-node tetrahedral elements.

source

FinEtoolsDeforLinear.FEMMDeforSurfaceDampingModule.FEMMDeforSurfaceDamping — Type.

FEMMDeforSurfaceDamping{S<:AbstractFESet, F<:Function} <: AbstractFEMM

Type for surface damping model.

source

FinEtoolsDeforLinear.FEMMDeforLinearNICEModule.AbstractFEMMDeforLinearNICE — Type.

AbstractFEMMDeforLinearNICE <: AbstractFEMMDeforLinear

Abstract FEMM type for Nodally Integrated Continuum Elements (NICE).

source

FinEtoolsDeforLinear.FEMMDeforLinearNICEModule.FEMMDeforLinearNICEH8 — Type.

FEMMDeforLinearNICEH8{MR<:AbstractDeforModelRed, S<:FESetH8, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearNICE

FEMM type for Nodally Integrated Continuum Elements (NICE) based on the eight-node hexahedron.

source

FinEtoolsDeforLinear.FEMMDeforLinearNICEModule.FEMMDeforLinearNICET4 — Type.

FEMMDeforLinearNICET4{MR<:AbstractDeforModelRed, S<:FESetT4, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearNICE

FEMM type for Nodally Integrated Continuum Elements (NICE) based on the 4-node tetrahedron.

source

FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule.AbstractFEMMDeforLinearESNICE — Type.

AbstractFEMMDeforLinearESNICE <: AbstractFEMMDeforLinear

Abstract FEMM type for Nodally Integrated Continuum Elements (NICE) with energy-sampling stabilization.

source

FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule.FEMMDeforLinearESNICEH8 — Type.

FEMMDeforLinearESNICEH8{MR<:AbstractDeforModelRed, S<:FESetH8, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearESNICE

FEMM type for Nodally Integrated Continuum Elements (NICE) based on the 8-node hexahedron.

source

FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule.FEMMDeforLinearESNICET4 — Type.

FEMMDeforLinearESNICET4{MR<:AbstractDeforModelRed, S<:FESetT4, F<:Function, M<:AbstractMatDeforLinearElastic} <: AbstractFEMMDeforLinearESNICE

FEMM type for Nodally Integrated Continuum Elements (NICE) based on the 4-node tetrahedron.

source

Material models

Material for deformation, base functionality

FinEtoolsDeforLinear.MatDeforModule.AbstractMatDefor — Type.

AbstractMatDefor

Abstract type that represents deformable materials.

source

Material models for elasticity

FinEtoolsDeforLinear.MatDeforLinearElasticModule.AbstractMatDeforLinearElastic — Type.

AbstractMatDeforLinearElastic <: AbstractMatDefor

Abstract Linear Elasticity material.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Type.

MatDeforElastIso{MR<:AbstractDeforModelRed, MTAN<:Function, MUPD<:Function, MTHS<:Function} <: AbstractMatDeforLinearElastic

Linear isotropic elasticity material.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{DeforModelRed1D}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 1D models.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{DeforModelRed2DAxisymm}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D axially symmetric models.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{DeforModelRed2DStrain}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D plane strain models.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{DeforModelRed2DStress}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D plane stress models.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{DeforModelRed3D}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 3D models.

source

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method.

MatDeforElastIso(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create isotropic elastic material with default mass density and thermal expansion.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Type.

MatDeforElastOrtho{MR<:AbstractDeforModelRed,  MTAN<:Function, MUPD<:Function, MTHS<:Function} <: AbstractMatDeforLinearElastic

Linear orthotropic elasticity material.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{DeforModelRed2DAxisymm}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D axially symmetric models.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{DeforModelRed2DStrain}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D plane strain models.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{DeforModelRed2DStress}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D plane stress models.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{DeforModelRed3D}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 3D models.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{MR}, E1::FFlt, E2::FFlt, E3::FFlt,
nu12::FFlt, nu13::FFlt, nu23::FFlt,
G12::FFlt, G13::FFlt, G23::FFlt) where {MR}

Create elastic orthotropic material.

Convenience version with only the specification of the elastic properties.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{MR}, mass_density::FFlt,  E::FFlt, nu::FFlt, CTE::FFlt) where {MR}

Create elastic orthotropic material which is really isotropic.

Convenience version with only the specification of the elastic and thermal expansion properties.

source

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method.

MatDeforElastOrtho(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create elastic orthotropic material which is really isotropic.

Convenience version with only the specification of the elastic properties.

source