Types

FEM machines

Linear deformation

Modules = [FinEtools, FinEtoolsDeforLinear.DeforModelRedModule, FinEtoolsDeforLinear.FEMMDeforLinearBaseModule, FinEtoolsDeforLinear.FEMMDeforLinearModule, FinEtoolsDeforLinear.FEMMDeforWinklerModule, FinEtoolsDeforLinear.FEMMDeforLinearMSModule, FinEtoolsDeforLinear.FEMMDeforSurfaceDampingModule, FinEtoolsDeforLinear.FEMMDeforLinearNICEModule, FinEtoolsDeforLinear.FEMMDeforLinearESNICEModule]
Private = true
Order = [:type]

Material models

Material for deformation, base functionality

FinEtoolsDeforLinear.MatDeforModule.AbstractMatDefor — Type

AbstractMatDefor

Abstract type that represents deformable materials.

Material models for elasticity

FinEtoolsDeforLinear.MatDeforLinearElasticModule.AbstractMatDeforLinearElastic — Type

AbstractMatDeforLinearElastic <: AbstractMatDefor

Abstract Linear Elasticity material.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Type

MatDeforElastIso{MR<:AbstractDeforModelRed, MTAN<:Function, MUPD<:Function, MTHS<:Function} <: AbstractMatDeforLinearElastic

Linear isotropic elasticity material.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed1DStrain}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 1D models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed1DStress}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 1D models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed2DAxisymm}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D axially symmetric models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed2DStrain}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D plane strain models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed2DStress}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 2D plane stress models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{DeforModelRed3D}, mass_density::FFlt, E::FFlt, nu::FFlt, CTE::FFlt)

Create elastic isotropic material for 3D models.

FinEtoolsDeforLinear.MatDeforElastIsoModule.MatDeforElastIso — Method

MatDeforElastIso(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create isotropic elastic material with default mass density and thermal expansion.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Type

MatDeforElastOrtho{MR<:AbstractDeforModelRed,  MTAN<:Function, MUPD<:Function, MTHS<:Function} <: AbstractMatDeforLinearElastic

Linear orthotropic elasticity material.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{DeforModelRed2DAxisymm}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D axially symmetric models.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{DeforModelRed2DStrain}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D plane strain models.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{DeforModelRed2DStress}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 2D plane stress models.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{DeforModelRed3D}, mass_density::FFlt,
	E1::FFlt, E2::FFlt, E3::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt,
	G12::FFlt, G13::FFlt, G23::FFlt, CTE1::FFlt, CTE2::FFlt, CTE3::FFlt)

Create elastic orthotropic material for 3D models.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create elastic orthotropic material which is really isotropic.

Convenience version with only the specification of the elastic properties.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{MR}, mass_density::FFlt,  E::FFlt, nu::FFlt, CTE::FFlt) where {MR}

Create elastic orthotropic material which is really isotropic.

Convenience version with only the specification of the elastic and thermal expansion properties.

FinEtoolsDeforLinear.MatDeforElastOrthoModule.MatDeforElastOrtho — Method

MatDeforElastOrtho(mr::Type{MR}, E1::FFlt, E2::FFlt, E3::FFlt,
nu12::FFlt, nu13::FFlt, nu23::FFlt,
G12::FFlt, G13::FFlt, G23::FFlt) where {MR}

Create elastic orthotropic material.

Convenience version with only the specification of the elastic properties.